just find all unresolveables
I asked myself the question, "What if we just evaluate every single
dependency in the first round instead of only the ones that we affected
with our ruleset?"

This question is mainly relevant for thinly populated architectures like
`freebsd-64`, where we hardly remove any packages since there are few to
begin with, and as a result we end up leaving most of the `noarch`
packages untouched even though they are not practically useable there.

It turns out that this isn't too expensive and seems to work better, so
for now, I am keeping it this way.